+++ plugins/hosting/install.php
@@ -4,7 +4,7 @@
 rex_sql_table::get(rex::getTable('project_manager_domain_hosting'))
     //->ensurePrimaryIdColumn()
     ->ensureColumn(new rex_sql_column('domain', 'varchar(255)', false, null, ''))->setPrimaryKey('domain')
-    ->ensureColumn(new rex_sql_column('raw', 'text', true))
+    ->ensureColumn(new rex_sql_column('raw', 'longtext', true))
     ->ensureColumn(new rex_sql_column('ip', 'text'))
     ->ensureColumn(new rex_sql_column('createdate', 'timestamp', false, '0000-00-00 00:00:00', 'on update CURRENT_TIMESTAMP'))
     ->ensureColumn(new rex_sql_column('status', 'text'))